refactor: Replace DefaultSyncObserver with FakeSyncObserver in tests and update dependencies
