+++ core/util/testing/build.gradle.kts
@@ -7,7 +7,6 @@ kotlin {
         commonMain {
             dependencies {
                 api(projects.core.appconfig.api)
-                api(projects.core.syncstate.api)
                 api(projects.core.util.api)
                 implementation(projects.core.base)
                 implementation(libs.coroutines.core)
