+++ features/discover/presenter/build.gradle.kts
@@ -59,6 +59,7 @@ kotlin {
                 implementation(projects.data.trendingshows.testing)
                 implementation(projects.data.upcomingshows.testing)
                 implementation(projects.data.user.testing)
+                implementation(projects.core.syncstate.testing)
                 implementation(projects.core.util.implementation)
                 implementation(projects.core.util.testing)
                 implementation(projects.i18n.testing)
